refactor: extract date formatting utilities to shared utils
- Move formatDate() and formatTime() to lib/utils.ts
- Remove duplicate code from user-detail and post-detail components
- Import shared utilities to maintain DRY principle
