+++ src/components/post/post-detail.tsx
@@ -1,32 +1,10 @@
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Card, CardContent, CardHeader } from "@/components/ui/card"
+import { formatDate, formatTime } from "@/lib/utils"
 import { Post } from "@/types/post"
 import { User } from "@/types/user"
 
 import { BookmarkButton } from "./bookmark-button"
-
-/**
- * Formats a date to a consistent format (YYYY-MM-DD)
- * This ensures server and client render the same output
- */
-function formatDate(dateInput: string | Date): string {
-  const date = dateInput instanceof Date ? dateInput : new Date(dateInput)
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, "0")
-  const day = String(date.getDate()).padStart(2, "0")
-  return `${year}-${month}-${day}`
-}
-
-/**
- * Formats a time to a consistent format (HH:MM)
- * This ensures server and client render the same output
- */
-function formatTime(dateInput: string | Date): string {
-  const date = dateInput instanceof Date ? dateInput : new Date(dateInput)
-  const hours = String(date.getHours()).padStart(2, "0")
-  const minutes = String(date.getMinutes()).padStart(2, "0")
-  return `${hours}:${minutes}`
-}
 
 interface PostDetailProps {
   post: Post
