+++ src/components/user/user-detail.tsx
@@ -2,6 +2,7 @@ import { memo } from "react"
 
 import { FollowButton } from "@/components/user/follow-button"
 import { UserAvatar } from "@/components/user/user-avatar"
+import { formatDate } from "@/lib/utils"
 import type { UserComponentProps } from "@/types/components"
 
 interface UserDetailProps extends UserComponentProps {
@@ -10,18 +11,6 @@ interface UserDetailProps extends UserComponentProps {
   isFollowing?: boolean
   onFollowToggle?: () => void
   showFollowButton?: boolean
-}
-
-/**
- * Formats a date string to a consistent format (YYYY-MM-DD)
- * This ensures server and client render the same output
- */
-function formatDate(dateInput: string | Date): string {
-  const date = dateInput instanceof Date ? dateInput : new Date(dateInput)
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, "0")
-  const day = String(date.getDate()).padStart(2, "0")
-  return `${year}-${month}-${day}`
 }
 
 /**
